+++ src/modules/delete_query_builder/delete_query_builder.rs
@@ -3,6 +3,7 @@ use proc_macro2::Ident;
 use quote::quote;
 use syn::DeriveInput;
 
+use crate::modules::query_builder_shared as shared;
 use crate::common::utils::{
     get_dbset_name, get_inner_option_type, get_key_fields, get_struct_name, get_table_name,
     get_unique_fields,
